+++ FlowRunner/Helpers/FlowHelper.cs
@@ -255,7 +255,7 @@ public class FlowHelper
         var instance = CreateFlowElementInstance(logger, part, nt, variables);
         if (instance == null)
             return instance;
-        if (instance.LicenseLevel > runner.runInstance.Config.LicenseLevel)
+        if ((int)instance.LicenseLevel > (int)runner.runInstance.Config.LicenseLevel)
             return Result<Node>.Fail(
                 $"The flow element {instance.GetType().Name} requires a {instance.LicenseLevel} license.");
         return instance;
